HUD is now on top
HUD's always on top.  Buttons should always show whichever abilities are
present (so long as they are in order from left to right [a q and e
can't be shown unless there's a  w too]).  Started work on the ability
description display.  A brief description of what I'm going to do can be
found at the bottom of the ShowTitle script in the HUD folder.